Updating getForces() to account for reference pressure (#193)
* Updating getForces() to find reference pressure.

Description: Fixing incorrect implementation of getForces routine.
Now it finds the reference pressure of the simulation and subtracts
it from the pressure at a certain face. Otherwise, the force at
each face / node is significantly overestimated.

* Adding options dictionary for FSI cases and updating getForces()

Description: To be explicit about the reference pressure used for
a case, I added an options dictionary for FSI cases that includes
a parameter for pRef. This value is used in the getForces()
function to calculate hte force due to pressure.